Add themes wrapper
