+++ public/themes/dark_mono.css
@@ -1,0 +1,7 @@
+:root {
+  --background: 309 19% 7%;
+  --secondary: 0 0% 80%;
+  --foreground: 0 0% 86%;
+  --tertiary: 0 0% 77%;
+  --hover: 0 0% 100%;
+}
