+++ public/themes/template.css
@@ -1,0 +1,7 @@
+:root {
+  --background: #000;
+  --secondary: #000;
+  --foreground: #000;
+  --tertiary: #000;
+  --hover: #000;
+}
